requirements.txt
flask~=3.1.1
flask-login~=0.6.3
python-dotenv~=1.1.0
wtforms~=3.2.1
flask-cors~=6.0.0